Refactor: move away from useEffect onChange (#478)
In Dashboard and Explore views this pattern was fine. However, Alerting
uses a different approach to state management which made this pattern
result in an infinite loop of state updates. Instead of having a
useEffect that computes new state and fires onChange with that state, I
wrapped each state change in it's own change handler to avoid the
infinite loop presented by useEffect + onChange.