fix: migrate to @grafana/llm package for OpenAI functionality
The @grafana/experimental package is being deprecated so development
of LLM-related functionality has been moved to a separate package.

The API is exactly the same so this _should_ just work.

I've had to fix the resolution of some Grafana packages which were
otherwise being duplicated because the @grafana/llm package has quite
a loose dependency range; without this, we had both version 10.4.1 and
11.3.0 of the various Grafana packages (and all of their
subdependencies) which was causing tests to fail.

Fixes #1051.